Reorganize crafting code
