+++ emergence_lib/src/water/roots.rs
@@ -7,7 +7,7 @@ use serde::{Deserialize, Serialize};
 
 use crate::{
     asset_management::manifest::Id,
-    crafting::components::{CraftingState, InputInventory},
+    crafting::inventories::{CraftingState, InputInventory},
     items::{item_manifest::ItemManifest, ItemCount},
     simulation::geometry::{Height, MapGeometry, TilePos, Volume},
     structures::structure_manifest::{Structure, StructureManifest},
